% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_Infer.R
\name{Infer}
\alias{Infer}
\title{Infer}
\usage{
Infer(data, prior = "Perk")
}
\arguments{
\item{data:}{A data.frame with the known data where the first column is the
assigned period and successive columns are the ordered types}

\item{prior:}{Prior to be selected. Options are "Haldane", "Jeffreys", "Perk",
"BayesLaplace","Hierarchical". Default is "Perk".}
}
\value{
A list where each element is the period. Each element/period is a data frame
with the curve for each type
}
\description{
It provides the theoretical probability distribution of the
types given the control (or any) dataset and a non-informative prior
}
