% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_Phase.pred.R
\name{Phase.pred}
\alias{Phase.pred}
\title{Phase.pred}
\usage{
Phase.pred(dated, estimate, prior = "Perk", site)
}
\arguments{
\item{dated:}{A data.frame with the known data where the first column is the
assigned period and successive columns are the ordered types}

\item{estimate:}{Data on which we want to estimate. Rownames are the names of
the sites and are used to compute the probability, so they must be correct.
Columns are number of each type per site}

\item{prior:}{Prior to be selected. Options are "Haldane", "Jeffreys", "Perk",
"BayesLaplace","Hierarchical". Default is "Perk".}

\item{site:}{The site on which we want to estimate. It must be spelled as in
the rownames}
}
\value{
a vector with the prediction for the selected site where each position
is the probability per period
}
\description{
It performs the Dirichlet process and returns the chronological
probability of an assemblage or site, given the control dataset
}
